"Rename weather module using PEP8"
