+++ reskit/weather/__init__.py
@@ -1,10 +1,10 @@
 from reskit.weather import util
 
-from reskit.weather.NCSource import NCSource
+from reskit.weather.nc_source import NCSource
 
-from reskit.weather.MerraSource.MerraSource import MerraSource
+from reskit.weather.merra_source.merra_source import MerraSource
 from reskit.weather.Era5Source.Era5Source import Era5Source
-from reskit.weather.SarahSource import SarahSource
+from reskit.weather.sarah_source import SarahSource
 from reskit.weather.GSA_mean.GSAmeanSource import GSAmeanSource
 from reskit.weather.GWA_mean.GWAmeanSource import GWAmeanSource
 from reskit.weather.IconlamSource.IconlamSource import IconlamSource
