+++ reskit/weather/merra_source/__init__.py
@@ -1,0 +1,1 @@
+from reskit.weather.merra_source.merra_source import MerraSource
