+++ reskit/weather/merra_source/data/gwa100_mean_over_merra.tif
