+++ reskit/weather/merra_source/data/merra_average_windspeed_50m.tif
